Adds manual requests installation for Travis
This change is necessary because pip does not detect that the new
duckduckgo fork requires requests, and attempts to install it before
requests is available (even though it's last in the requirements).

To be fixed by either using or writing another duckduckgo module (we
only use instant answer anyway) or doing some other trick which lets pip
install requests first. (It works for now, but is not entirely nice)